BUG 1 While debugging, I realized that the problem wasn’t from the winnings as the receiveWinning method gets the right winnings. The problem is that when a bet is placed, the bet is removed from the balance but when the Player wins, only the winnings is added, the removed bet is not added back. To correct the this, I added the 5(bet) to the balance so that when a Player wins, his bet returns to him
BUG 2
The problem is that the Player is supposed to be able to place a bet
when he still has a balance of 5 but is unable to. To fix this I changed
this method to return (balance-amount >=limit) unlike before when it
returned balance-amount >limit. 
Assuming balance=5 and player wants to bet 5 as usual, 5-5=0>0 returns
false so the Player cannot bet even when he has enough money to bet.
Just make it greater than or equals to and the program functions
correctly because 5-5>=0 returns true which allow the player to bet the
remaining 5.